convert line model to double precision, to support continuous dragging

+++ simulations-java/simulations/line-graphing/src/edu/colorado/phet/linegraphing/intro/view/InteractiveLineGraphNode.java
@@ -118,7 +118,7 @@ public class InteractiveLineGraphNode extends LineGraphNode {
 
 
         // move the manipulators
-        final int y = line.rise + line.intercept;
+        final double y = line.rise + line.intercept;
         double x;
         if ( line.run == 0 ) {
             x = 0;
